+++ e2e/autoscript-apps/src/authn-oauth/autoscript.ts
@@ -14,6 +14,7 @@ import { from } from 'rxjs';
 
 function autoscript() {
   const delay = 0;
+  const tokenExpiredDelay = 2000;
 
   const url = new URL(window.location.href);
   const amUrl = url.searchParams.get('amUrl') || 'https://auth.example.com:9443/am';
@@ -23,6 +24,10 @@ function autoscript() {
   const un = url.searchParams.get('un') || 'sdkuser';
   const pw = url.searchParams.get('pw') || 'password';
   const tree = url.searchParams.get('tree') || 'UsernamePassword';
+  const oauthThreshold = url.searchParams.get('oauthThreshold');
+  const resourceOrigin =
+    url.searchParams.get('resourceOrigin') || 'https://api.example.com:9443/resource';
+  const igUrl = url.searchParams.get('igUrl'); // only use when testing against IG on different host
 
   console.log('Configure the SDK');
   forgerock.Config.set({
@@ -47,6 +52,7 @@ function autoscript() {
     serverConfig: {
       baseUrl: amUrl,
     },
+    oauthThreshold: oauthThreshold,
   });
 
   try {