+++ e2e/mock-api/src/app/wait.js
@@ -1,0 +1,15 @@
+/*
+ * @forgerock/javascript-sdk
+ *
+ * wait.js
+ *
+ * Copyright (c) 2020 ForgeRock. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+const delay = 0;
+
+export default function wait(req, res, next) {
+  setTimeout(() => next(), delay);
+}