+++ samples/angular-todo/src/app/app-routing.module.ts
@@ -1,0 +1,36 @@
+/*
+ * angular-todo-prototype
+ *
+ * app-routing.module.ts
+ *
+ * Copyright (c) 2021 ForgeRock. All rights reserved.
+ * This software may be modified and distributed under the terms
+ * of the MIT license. See the LICENSE file for details.
+ */
+
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from './views/home/home.component';
+import { LoginComponent } from './views/login/login.component';
+import { TodosComponent } from './views/todos/todos.component';
+import { AuthGuard } from './auth/auth.guard';
+import { RegisterComponent } from './views/register/register.component';
+import { LogoutComponent } from './features/logout/logout.component';
+
+const routes: Routes = [
+  { path: 'home', component: HomeComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'register', component: RegisterComponent },
+  { path: 'todos', canActivate: [AuthGuard], component: TodosComponent },
+  { path: 'logout', component: LogoutComponent },
+  { path: '', redirectTo: 'home', pathMatch: 'full' },
+];
+
+/**
+ * Defines the routes and auth guards for the application
+ */
+@NgModule({
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+})
+export class AppRoutingModule {}